backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pandas==2.2.2
joblib==1.4.2
scikit-learn==1.4.2
xgboost==2.0.3
reportlab==4.2.0
qrcode[pil]==7.4.2
Pillow==10.3.0
web3==6.19.0
earthengine-api==0.1.398
requests==2.32.3
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
